64 bit seed.

+++ src/fixed_rand.cpp
@@ -4,7 +4,7 @@ namespace dux {
 
 FRandGenerator::FRandGenerator(int32_t seed) : rng_(seed) {}
 
-void FRandGenerator::SetSeed(int32_t seed) {
+void FRandGenerator::SetSeed(int64_t seed) {
   rng_.seed(seed);
 }
